+++ CHANGELOG.md
@@ -19,6 +19,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * implement `Clone` for all types
 * implement `Copy`, `Eq`, `Ord` and `Hash` on `Account` and `Currency`
 
+### Documentation
+
+* Write documentation for `Transaction` type
+
 
 ## [1.0.0-alpha.1] - 2023-05-21